extract printCharacterList to a file
